Add image filter and command line parametersdddddddd
Add the ability to apply a filter to the rain so
that the rain effect is shown with the shape of the
provided image.
Also add various command line parameters to tweak
the effect (e.g. offset the image mask from top left
corner or flip black/white)

Use minimist to handle command line parameters,
reviewing the previous ones too (direction and source
file)